Create web-scraper
This is a web scraper for my job search. The results from this are a list of jobs titles, companies, and location pulled from Monster.com. I've since built scrapers for the websites of companies for which I'm particularly interested in working. Dissecting the HTML of various sites and adapting/writing the Python script accordingly has been challenging, but educational.